add favorites and recently deleted tabs + add user info with each card

+++ convex/files.ts
@@ -15,17 +15,25 @@ export const generateUploadUrl = mutation(async (ctx) => {
   return await ctx.storage.generateUploadUrl();
 });
 
-async function hasAccessToOrg(
-  ctx: QueryCtx | MutationCtx,
-  tokenIdentifier: string,
-  orgId: string
-) {
-  const user = await getUserByToken(ctx, tokenIdentifier);
+async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
+  const identity = await ctx.auth.getUserIdentity();
+
+  if (!identity) {
+    throw new ConvexError(
+      "Unauthorized, you must be signed in to create a file."
+    );
+  }
+
+  const user = await getUserByToken(ctx, identity.tokenIdentifier);
 
   const isOrgMember = user.orgIds.includes(orgId);
   const isPersonalOrg = user.tokenIdentifier.includes(orgId);
 
-  return isOrgMember || isPersonalOrg;
+  return {
+    hasAccess: isOrgMember || isPersonalOrg,
+    user,
+    tokenIdentifier: identity.tokenIdentifier,
+  };
 }
 
 export const createFile = mutation({
@@ -36,19 +44,7 @@ export const createFile = mutation({
     orgId: v.string(),
   },
   async handler(ctx, args) {
-    const identity = await ctx.auth.getUserIdentity();
-
-    if (!identity) {
-      throw new ConvexError(
-        "Unauthorized, you must be signed in to create a file."
-      );
-    }
-
-    const hasAccess = await hasAccessToOrg(
-      ctx,
-      identity.tokenIdentifier,
-      args.orgId
-    );
+    const { hasAccess, user } = await hasAccessToOrg(ctx, args.orgId);
 
     if (!hasAccess) {
       throw new ConvexError(
@@ -60,6 +56,7 @@ export const createFile = mutation({
       name: args.name,
       type: args.type,
       orgId: args.orgId,
+      userId: user._id,
       fileId: args.fileId,
     });
   },
@@ -70,17 +67,13 @@ export const getFiles = query({
     orgId: v.string(),
     searchQuery: v.optional(v.string()),
     favorites: v.optional(v.boolean()),
+    deleted: v.optional(v.boolean()),
+    filter: v.optional(FileType),
   },
+
   async handler(ctx, args) {
-    const identity = await ctx.auth.getUserIdentity();
-
-    if (!identity) {
-      return [];
-    }
-
-    const hasAccess = await hasAccessToOrg(
+    const { hasAccess, tokenIdentifier } = await hasAccessToOrg(
       ctx,
-      identity.tokenIdentifier,
       args.orgId
     );
 
@@ -95,6 +88,8 @@ export const getFiles = query({
 
     const query = args.searchQuery;
     const favortes = args.favorites;
+    const deleted = args.deleted;
+    const filter = args.filter;
 
     if (query) {
       files = files.filter((file) =>
@@ -102,46 +97,59 @@ export const getFiles = query({
       );
     }
 
-    if (favortes) {
-      const user = await getUserByToken(ctx, identity.tokenIdentifier);
-      const favorites = await ctx.db
-        .query("favoriteFiles")
-        .withIndex("byUserId_byOrgId_byFileId", (q) =>
-          q.eq("userId", user._id).eq("orgId", args.orgId)
-        )
-        .collect();
+    const user = await getUserByToken(ctx, tokenIdentifier);
 
+    const deletedFiles = await ctx.db
+      .query("deletedFiles")
+      .withIndex("byUserId_byOrgId_byFileId", (q) =>
+        q.eq("userId", user._id).eq("orgId", args.orgId)
+      )
+      .collect();
+
+    if (deleted) {
       files = files.filter((file) =>
-        favorites.some((f) => f.fileId === file._id)
+        deletedFiles.some((f) => f.fileId === file._id)
       );
+    } else {
+      if (favortes) {
+        const favorites = await ctx.db
+          .query("favoriteFiles")
+          .withIndex("byUserId_byOrgId_byFileId", (q) =>
+            q.eq("userId", user._id).eq("orgId", args.orgId)
+          )
+          .collect();
+
+        files = files.filter((file) =>
+          favorites.some((f) => f.fileId === file._id)
+        );
+      }
+
+      files = files.filter(
+        (file) => !deletedFiles.some((f) => f.fileId === file._id)
+      );
+    }
+
+    if (filter) {
+      files = files.filter((file) => file.type === filter);
     }
 
     return files;
   },
 });
 
-export const deleteFile = mutation({
+export const toggleDeleteFile = mutation({
   args: {
     fileId: v.id("files"),
   },
   async handler(ctx, args) {
-    const identity = await ctx.auth.getUserIdentity();
-
-    if (!identity) {
-      throw new ConvexError(
-        "Unauthorized, you must be signed in to delete a file."
-      );
-    }
-
     const file = await ctx.db.get(args.fileId);
 
     if (!file) {
       throw new ConvexError("File not found.");
     }
 
-    const hasAccess = await hasAccessToOrg(
+    const { hasAccess, tokenIdentifier } = await hasAccessToOrg(
       ctx,
-      identity.tokenIdentifier,
       file.orgId
     );
 
@@ -151,6 +159,47 @@ export const deleteFile = mutation({
       );
     }
 
+    const user = await getUserByToken(ctx, tokenIdentifier);
+
+    const deletedFile = await ctx.db
+      .query("deletedFiles")
+      .withIndex("byUserId_byOrgId_byFileId", (q) =>
+        q.eq("userId", user._id).eq("orgId", file.orgId).eq("fileId", file._id)
+      )
+      .first();
+
+    if (!deletedFile) {
+      await ctx.db.insert("deletedFiles", {
+        fileId: file._id,
+        userId: user._id,
+        orgId: file.orgId,
+      });
+    } else {
+      await ctx.db.delete(deletedFile._id);
+    }
+  },
+});
+
+export const perminantlyDeleteFile = mutation({
+  args: {
+    fileId: v.id("files"),
+  },
+  async handler(ctx, args) {
+    const file = await ctx.db.get(args.fileId);
+
+    if (!file) {
+      throw new ConvexError("File not found.");
+    }
+
+    const { hasAccess } = await hasAccessToOrg(ctx, file.orgId);
+
+    if (!hasAccess) {
+      throw new ConvexError(
+        "Unauthorized, you do not have access to this org."
+      );
+    }
+
+    await ctx.storage.delete(file.fileId);
     await ctx.db.delete(args.fileId);
   },
 });
@@ -160,21 +209,14 @@ export const toggleFavorite = mutation({
     fileId: v.id("files"),
   },
   async handler(ctx, args) {
-    const identity = await ctx.auth.getUserIdentity();
-
-    if (!identity) {
-      throw new ConvexError("Unauthorized, you must be signed in.");
-    }
-
     const file = await ctx.db.get(args.fileId);
 
     if (!file) {
       throw new ConvexError("File not found.");
     }
 
-    const hasAccess = await hasAccessToOrg(
+    const { hasAccess, tokenIdentifier } = await hasAccessToOrg(
       ctx,
-      identity.tokenIdentifier,
       file.orgId
     );
 
@@ -184,7 +226,7 @@ export const toggleFavorite = mutation({
       );
     }
 
-    const user = await getUserByToken(ctx, identity.tokenIdentifier);
+    const user = await getUserByToken(ctx, tokenIdentifier);
 
     const favorite = await ctx.db
       .query("favoriteFiles")
@@ -202,5 +244,49 @@ export const toggleFavorite = mutation({
     } else {
       await ctx.db.delete(favorite._id);
     }
+  },
+});
+
+export const getAllFavorites = query({
+  args: {
+    orgId: v.string(),
+  },
+  async handler(ctx, args) {
+    const { hasAccess, user } = await hasAccessToOrg(ctx, args.orgId);
+
+    if (!hasAccess) {
+      return [];
+    }
+
+    const favorites = await ctx.db
+      .query("favoriteFiles")
+      .withIndex("byUserId_byOrgId_byFileId", (q) =>
+        q.eq("userId", user._id).eq("orgId", args.orgId)
+      )
+      .collect();
+
+    return favorites;
+  },
+});
+
+export const getDeletedFiles = query({
+  args: {
+    orgId: v.string(),
+  },
+  async handler(ctx, args) {
+    const { hasAccess, user } = await hasAccessToOrg(ctx, args.orgId);
+
+    if (!hasAccess) {
+      return [];
+    }
+
+    const deletedFiles = await ctx.db
+      .query("deletedFiles")
+      .withIndex("byUserId_byOrgId_byFileId", (q) =>
+        q.eq("userId", user._id).eq("orgId", args.orgId)
+      )
+      .collect();
+
+    return deletedFiles;
   },
 });
